requirements.txt
PyQt5
ultralytics
opencv-python
kagglehub
PyYAML
pytest
pytest-qt

